+++ lib/core/decision_making_enums.dart
@@ -24,4 +24,5 @@ enum SawProcessingIsolateCommand implements DecisionProcessingIsolateCommand {
 enum TopsisProcessingIsolateCommand implements DecisionProcessingIsolateCommand {
   generateMatrix,
   normalizeMatrix,
+  normalizeEuclidean,
 }