+++ lib/feature/topsis/domain/entities/topsis_raw_matrix.dart
@@ -2,12 +2,22 @@ import 'package:flutter_decision_making/core/shared/entity/weighted_decision_cri
 import 'package:flutter_decision_making/core/shared/entity/weighted_decision_matrix.dart';
 
 /// CALCULATE SESSION
-class TopsisSession {
+class TopsisRawMatrix {
   final List<WeightedDecisionCriteria> criterias;
   final List<WeightedDecisionMatrix> matrixs;
 
-  TopsisSession({
+  TopsisRawMatrix({
     required this.criterias,
     required this.matrixs,
   });
+
+  TopsisRawMatrix copyWith({
+    List<WeightedDecisionCriteria>? criterias,
+    List<WeightedDecisionMatrix>? matrixs,
+  }) {
+    return TopsisRawMatrix(
+      criterias: criterias ?? this.criterias,
+      matrixs: matrixs ?? this.matrixs,
+    );
+  }
 }