+++ lib/feature/topsis/domain/usecase/generate_topsis_matrix_usecase.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_decision_making/core/shared/entity/weighted_decision_alternative.dart';
 import 'package:flutter_decision_making/core/shared/entity/weighted_decision_criteria.dart';
-import 'package:flutter_decision_making/core/shared/entity/weighted_decision_matrix.dart';
+import 'package:flutter_decision_making/feature/topsis/domain/entities/topsis_raw_matrix.dart';
 import 'package:flutter_decision_making/feature/topsis/domain/repository/topsis_repository.dart';
 
 class GenerateTopsisMatrixUsecase {
@@ -8,10 +8,12 @@ class GenerateTopsisMatrixUsecase {
 
   GenerateTopsisMatrixUsecase(this._repository);
 
-  Future<List<WeightedDecisionMatrix>> execute({
+  Future<TopsisRawMatrix> execute({
     required List<WeightedDecisionAlternative> listAlternative,
     required List<WeightedDecisionCriteria> listCriteria,
   }) async =>
-      _repository.generateSawMatrix(
-          listAlternative: listAlternative, listCriteria: listCriteria);
+      await _repository.generateTopsisMatrix(
+        listAlternative: listAlternative,
+        listCriteria: listCriteria,
+      );
 }